requirements.txt
Cython==0.21.1
ecdsa==0.13
Fabric==1.10.1
gnureadline==6.3.3
ipython==2.3.1
matplotlib==1.4.2
mock==1.0.1
nose==1.3.4
numpy==1.9.2
paramiko==1.15.2
Pillow==2.8.1
pycrypto==2.6.1
pyparsing==2.0.3
python-dateutil==2.3
pytz==2014.10
six==1.8.0
v4l2==0.2
virtualenv==1.11.6
wxPython==3.0.2.0
wxPython-common==3.0.2.0
